requirements.txt
# Core dependencies
pandas==2.1.4
numpy==1.26.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0

# Data processing
scikit-learn==1.3.2
scipy==1.11.4

# Database
SQLAlchemy==2.0.23
alembic==1.13.1

# API (for later phases)
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# Bayesian modeling
pymc==5.10.0
arviz==0.17.0

# Visualization (for testing)
plotly==5.18.0
matplotlib==3.8.2

# Testing
pytest==7.4.3

